+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.6 (WIP)
 
 * Remove "chrome-extension:/" bit from error logs
+* Add `reuse_names` setting, which will reuse names once they all have been used once
 
 ## 0.1.5 (2018-03-26)